rename `datapoints` as `reads`
